# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Emailed API
  description: |
    The Emailed API provides programmatic access to the Emailed email
    infrastructure platform. Send transactional and bulk emails, manage
    domains, configure webhooks, and query analytics.
  version: 0.1.0
  contact:
    name: Emailed Support
    url: https://emailed.dev/support
    email: support@emailed.dev
  license:
    name: Proprietary
    url: https://emailed.dev/terms
servers:
  - url: https://api.emailed.dev
    description: Production
  - url: https://api.staging.emailed.dev
    description: Staging
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /v1/templates:
    post:
      operationId: createTemplate
      summary: Create an email template
      tags: [Templates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, subject]
              properties:
                name:
                  type: string
                subject:
                  type: string
                htmlBody:
                  type: string
                textBody:
                  type: string
                metadata:
                  type: object
      responses:
        "201":
          description: Template created
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      operationId: listTemplates
      summary: List templates
      tags: [Templates]
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/CursorParam"
        - name: name
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of templates
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/templates/{id}:
    get:
      operationId: getTemplate
      summary: Get template details
      tags: [Templates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Template details
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateTemplate
      summary: Update a template
      tags: [Templates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                subject:
                  type: string
                htmlBody:
                  type: string
                textBody:
                  type: string
                metadata:
                  type: object
      responses:
        "200":
          description: Template updated
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTemplate
      summary: Delete a template
      tags: [Templates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Template deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/templates/{id}/render:
    post:
      operationId: renderTemplate
      summary: Render a template with variables
      tags: [Templates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [variables]
              properties:
                variables:
                  type: object
      responses:
        "200":
          description: Rendered template
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      subject:
                        type: string
                      htmlBody:
                        type: string
                      textBody:
                        type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Pass a Bearer token in the Authorization header.
        Example: `Authorization: Bearer em_live_abc123...`
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Pass an API key in the X-API-Key header.
        Example: `X-API-Key: em_live_abc123...`
...
  parameters:
    CursorParam:
      name: cursor
      in: query
      description: Pagination cursor from a previous response
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return (1-100, default 20)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message, code]
          properties:
            type:
              type: string
              description: Error category
              example: validation_error
            message:
              type: string
              description: Human-readable error message
              example: "Invalid email address in 'to' field"
            code:
              type: string
              description: Machine-readable error code
              example: invalid_recipient
            context:
              type: object
              additionalProperties: true
              description: Additional error context
  responses:
    BadRequest:
      description: The request body or parameters are malformed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication credentials are missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...